@@ -4,44 +4,44 @@ repository: https://github.com/flutter/gallery
 version: 2.9.3+020903
 
 environment:
-  flutter: ^2.13.0-0 # Kept relatively close to master channel version
-  sdk: '>=2.12.0 <3.0.0'
+  flutter: ^3.1.0-0 # Kept relatively close to master channel version
+  sdk: '>=2.17.0 <3.0.0'
 
 dependencies:
   flutter:
     sdk: flutter
   flutter_localizations:
     sdk: flutter
-  adaptive_breakpoints: ^0.1.0
-  animations: ^2.0.0
-  collection: ^1.15.0
-  cupertino_icons: ^1.0.2
+  adaptive_breakpoints: ^0.1.1
+  animations: ^2.0.2
+  collection: ^1.16.0
+  cupertino_icons: ^1.0.4
   flutter_gallery_assets: ^1.0.2
-  flutter_localized_locales: ^2.0.1
+  flutter_localized_locales: ^2.0.3
   flutter_staggered_grid_view: ^0.6.1
-  google_fonts: ^2.0.0
+  google_fonts: ^2.3.2
   intl: ^0.17.0
-  meta: ^1.3.0
-  package_info_plus: ^1.0.1
-  provider: ^6.0.0
+  meta: ^1.7.0
+  package_info_plus: ^1.4.2
+  provider: ^6.0.2
   rally_assets: ^3.0.1
   scoped_model: ^2.0.0-nullsafety.0
   shrine_images: ^2.0.1
-  url_launcher: ^6.0.3
-  vector_math: ^2.1.0
-  shared_preferences: ^2.0.9
+  url_launcher: ^6.1.2
+  vector_math: ^2.1.2
+  shared_preferences: ^2.0.15
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
   flutter_driver:
     sdk: flutter
-  args: ^2.1.0
-  flutter_lints: ^2.0.0
-  grinder: ^0.9.0
-  path: ^1.8.0
+  args: ^2.3.1
+  flutter_lints: ^2.0.1
+  grinder: ^0.9.1
+  path: ^1.8.1
   string_scanner: ^1.1.0
-  test: ^1.16.8
+  test: ^1.21.1
   web_benchmarks: ^0.0.7
 
 flutter:
