@@ -330,12 +330,10 @@ class SettingsFeedback extends StatelessWidget {
       title: GalleryLocalizations.of(context)!.settingsFeedback,
       icon: Icons.feedback,
       onTap: () async {
-        const url = 'https://github.com/flutter/gallery/issues/new/choose/';
-        if (await canLaunch(url)) {
-          await launch(
-            url,
-            forceSafariVC: false,
-          );
+        final url =
+            Uri.parse('https://github.com/flutter/gallery/issues/new/choose/');
+        if (await canLaunchUrl(url)) {
+          await launchUrl(url);
         }
       },
     );
